api/pyproject.toml
[project]
name = "agri-backend"
version = "0.1.0"
description = "Agricultural prediction and management API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Django>=5.1.2",
    "djangorestframework>=3.15.0",
    "django-cors-headers>=4.0.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scikit-learn==1.5.1",
    "scipy>=1.13.0",
    "seaborn>=0.13.0",
    "matplotlib>=3.9.0",
    "tensorflow>=2.16.0",
    "whitenoise>=6.7.0",
    # Potential legacy/unused dependencies found in requirements.txt
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "transformers>=4.38.0",
    "torch>=2.2.0",
    "google-genai>=1.56.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
